requirements.txt
functions-framework==3.*
google-cloud-storage
google-cloud-pubsub
google-cloud-video-transcoder
google-cloud-videointelligence

